+++ public/ide/modules/ide/js/parts/nav/navController.js
@@ -314,7 +314,7 @@
                     curScope.project.version = window.ideVersion;
                     curScope.project.CANId = NavModalCANConfigService.getCANId();
                     var currentProject = curScope.project;
-                    // console.log(currentProject);
+                    console.log('currentProject',currentProject);
                     var thumb=_.cloneDeep(currentProject.pages[0].url);
                     scaleImg(thumb,['jpeg'],200,200,true, function (scaledThumb) {
                         _.forEach(currentProject.pages,function (_page) {
@@ -676,51 +676,89 @@
          */
 
         function generateDataFile(format){
-            generateData(format);
-            if (window){
-                if (window.spinner){
-                    window.spinner.setBackgroundColor('rgba(0,0,0,0.5)');
-                    window.spinner.show();
-                }
-                RenderSerive.renderProject(window.projectData,function () {
-                    toastr.info('生成成功');
-                    window.spinner&&window.spinner.hide();
-                },function () {
-                    toastr.info('生成失败');
-                    window.spinner&&window.spinner.hide();
-                });
-            }else{
-                saveProject(function () {
-                    showSpinner();
+            if(format=='local'){
+                
+                //console.log('keke',format);
+                var curScope = {};
+                ProjectService.getProjectCopyTo(curScope);
+                curScope.project.resourceList = ResourceService.getAllResource();
+
+                curScope.project.customTags = TagService.getAllCustomTags();
+                curScope.project.timerTags = TagService.getAllTimerTags();
+                curScope.project.timers = TagService.getTimerNum();
+                curScope.project.version = window.ideVersion;
+                curScope.project.CANId = NavModalCANConfigService.getCANId();
+                var currentProject = curScope.project;
+                //saveProject(function(){
+                    if (window.spinner){
+                        window.spinner.setBackgroundColor('rgba(0,0,0,0.5)');
+                        window.spinner.show();
+                    }
                     $http({
                         method:'POST',
-                        url:'/project/'+$scope.project.projectId+'/generate',
-                        data:{
-                            dataStructure:window.projectData
+                        url:'/project/'+$scope.project.projectId+'/generateLocalProject'
+                        //data:{currentProject:currentProject}
+                    })
+                    .success(function(data,status,xhr){
+                        //console.log(data);
+                        window.spinner&&window.spinner.hide();
+                        if(data=='ok'){
+                            toastr.info('生成本地版成功');
+                            window.location.href = '/project/'+$scope.project.projectId+'/downloadLocalProject'
+                        }else{
+                            toastr.error('生成失败')
                         }
                     })
-                        .success(function (data,status,xhr) {
-                            hideSpinner();
-                            if (data == 'ok'){
-                                toastr.info('生成成功');
-                                //download
-                                window.location.href = '/project/'+$scope.project.projectId+'/download'
-                            }else{
-                                console.log(data);
-                                toastr.info('生成失败')
+                    .error(function(err,status,xhr){
+                        console.log(err);
+                    })
+                //})
+            }else{
+                generateData(format);
+                if (window){
+                    if (window.spinner){
+                        window.spinner.setBackgroundColor('rgba(0,0,0,0.5)');
+                        window.spinner.show();
+                    }
+                    RenderSerive.renderProject(window.projectData,function () {
+                        toastr.info('生成成功');
+                        window.spinner&&window.spinner.hide();
+                    },function () {
+                        toastr.info('生成失败');
+                        window.spinner&&window.spinner.hide();
+                    });
+                }else{
+                    saveProject(function () {
+                        showSpinner();
+                        $http({
+                            method:'POST',
+                            url:'/project/'+$scope.project.projectId+'/generate',
+                            data:{
+                                dataStructure:window.projectData
                             }
-
-
                         })
-                        .error(function (err,status,xhr) {
-                            hideSpinner();
-                            console.log(err);
-                            toastr.info('生成失败')
+                            .success(function (data,status,xhr) {
+                                hideSpinner();
+                                if (data == 'ok'){
+                                    toastr.info('生成成功');
+                                    //download
+                                    window.location.href = '/project/'+$scope.project.projectId+'/download'
+                                }else{
+                                    console.log(data);
+                                    toastr.info('生成失败')
+                                }
 
-                        })
-                })
+
+                            })
+                            .error(function (err,status,xhr) {
+                                hideSpinner();
+                                console.log(err);
+                                toastr.info('生成失败')
+
+                            })
+                    })
+                }
             }
-
         }
 
         function generateData(format){
@@ -1004,6 +1042,7 @@
                     }
                 })
                 .error(function(data,status,xhr){
+                    toastr.error('导入失败');
                     console.log('导入失败',data);
                 });
             }
@@ -1065,6 +1104,13 @@ ide.controller('NavModalCtl',['$scope','$uibModalInstance',function ($scope,$uib
             name:'压缩'
         }
     ];
+    var localFormat = {
+        type:'local',
+        name:'本地'
+    }
+    if(!window.local){
+        $scope.formats[2] = localFormat;
+    };
     $scope.generateFormat = 'normal';
     $scope.ok = function () {
         $uibModalInstance.close({