+++ changelog.md
@@ -30,3 +30,14 @@ update:
     1. num bi-direction animation
 fix:
     1. font privilege
+    
+v1.8.5_test
+add:
+    1. export local project from web
+    2. CAN configure project for NW
+update:
+    1. CAN date structure
+    2. new draw datetime widget method
+    3. new draw num widget method
+fix:
+